Confirmation d'un candidat

+++ src/main/java/fr/ubo/spibackend/repositories/CandidatRepository.java
@@ -15,5 +15,5 @@ import java.util.List;
 public interface CandidatRepository extends JpaRepository<Candidat, String> {
 Candidat findByEmail(String email);
     List<Candidat> findByAnneeUniversitaireAndCodeFormation(String anneeUniversitaire,String codeFormation);
-    Candidat findBySelectionNoOrdre(BigInteger selectionNoOrdre);
+    Candidat findBySelectionNoOrdre(Integer selectionNoOrdre);
 }
